requirements.txt
aiohttp==3.8.5
aiosignal==1.3.1
apturl==0.5.2
asgiref==3.7.2
async-generator==1.10
async-timeout==4.0.3
attrs==23.1.0
autopep8==1.6.0
bcrypt==3.2.0
black==23.7.0
bleach==6.0.0
blinker==1.4
boto3==1.28.31
botocore==1.31.31
Brlapi==0.8.3
certifi==2023.7.22
cffi==1.15.0
chardet==4.0.0
charset-normalizer==3.2.0
click==8.1.7
colorama==0.4.4
command-not-found==0.3
coreapi==2.3.3
coreschema==0.0.4
coverage==6.2
cryptography==36.0.0
cupshelpers==1.0
dbus-python==1.2.18
defer==1.0.6
defusedxml==0.8.0rc2
distlib==0.3.4
distro==1.7.0
distro-info==1.1+ubuntu0.1
Django==4.2.4
django-cors-headers==4.2.0
django-elasticsearch-dsl==8.0
django-extensions==3.2.3
django-filter==23.2
django-health-check==3.17.0
django-js-asset==2.1.0
django-mptt==0.14.0
django-safedelete==1.3.2
django-storages==1.13.2
django-summernote==0.8.20.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.1.0
dnspython==2.4.2
docker==5.0.3
docker-compose==1.29.2
dockerpty==0.4.1
docopt==0.6.2
drf-yasg==1.21.7
duplicity==0.8.21
elastic-transport==8.4.1
elasticsearch==8.10.1
elasticsearch-dsl==8.9.0
Faker==10.0.0
fasteners==0.14.1
filelock==3.6.0
flake8==4.0.1
fluent-logger==0.10.0
frozenlist==1.4.0
future==0.18.2
gunicorn==21.2.0
h11==0.12.0
httplib2==0.20.2
idna==3.4
importlib-metadata==4.6.4
inflection==0.5.1
iniconfig==1.1.1
itypes==1.2.0
jeepney==0.7.1
Jinja2==3.1.2
jmespath==1.0.1
jsonschema==3.2.0
keyring==23.5.0
language-selector==0.1
launchpadlib==1.10.16
lazr.restfulclient==0.14.4
lazr.uri==1.0.6
lockfile==0.12.2
louis==3.20.0
macaroonbakery==1.3.1
Mako==1.2.4
MarkupSafe==2.1.3
mccabe==0.6.1
monotonic==1.6
more-itertools==8.10.0
msgpack==1.0.5
multidict==6.0.4
mypy==0.910
mypy-extensions==1.0.0
mysqlclient==1.4.6
netifaces==0.11.0
oauthlib==3.2.0
olefile==0.46
outcome==1.1.0
packaging==23.1
parameterized==0.8.1
paramiko==2.9.3
pathspec==0.11.2
pbr==5.8.0
pexpect==4.8.0
Pillow==8.4.0
platformdirs==3.10.0
pluggy==1.0.0
protobuf==3.12.4
psycopg2-binary==2.9.7
ptyprocess==0.7.0
py==1.11.0
pycairo==1.20.1
pycodestyle==2.8.0
pycparser==2.21
pycups==2.0.1
pyflakes==2.4.0
PyGObject==3.42.1
PyJWT==2.8.0
pymacaroons==0.13.0
pymongo==4.5.0
PyMySQL==1.1.0
PyNaCl==1.5.0
pyOpenSSL==21.0.0
pyparsing==3.0.6
pyRFC3339==1.1
pyrsistent==0.18.1
pytest==6.2.5
pytest-django==4.5.2
pytest-dotenv==0.5.2
pytest-watch==4.2.0
python-apt==2.4.0+ubuntu2
python-dateutil==2.8.2
python-debian==0.1.43+ubuntu1.1
python-decouple==3.8
python-dotenv==0.19.2
python3-openid==3.2.0
pytz==2023.3
pyxdg==0.27
PyYAML==6.0.1
reportlab==3.6.8
requests==2.31.0
requests-mock==1.11.0
requests-oauthlib==1.3.1
s3transfer==0.6.2
SecretStorage==3.3.1
selenium==4.1.0
six==1.16.0
sniffio==1.2.0
social-auth-app-django==5.4.0
social-auth-core==4.5.0
sortedcontainers==2.4.0
sqlparse==0.4.4
stevedore==3.5.0
systemd-python==234
text-unidecode==1.3
texttable==1.6.4
toml==0.10.2
trio==0.19.0
trio-websocket==0.9.2
typing_extensions==4.0.1
ubuntu-advantage-tools==8001
ubuntu-drivers-common==0.0.0
ufw==0.36.1
unattended-upgrades==0.1
Unidecode==1.3.6
uritemplate==4.1.1
urllib3==1.26.16
usb-creator==0.3.7
virtualenv==20.13.0+ds
virtualenv-clone==0.3.0
virtualenvwrapper==4.8.4
wadllib==1.3.6
watchdog==2.1.6
webencodings==0.5.1
websocket-client==1.2.3
whitenoise==5.3.0
wsproto==1.0.0
xdg==5
xkit==0.0.0
yarl==1.9.2
zipp==1.0.0
